CMakeLists.txt: enable out-of-tree build
The file created via the configure_file() function, will be
automatically put into the PROJECT_BINARY_DIR.

In the case of the in-tree build, PROJECT_BINARY_DIR is the same as
PROJECT_SOURCE_DIR. Hence, config.h is created in the same folder as
config.h.in.

In the case of the out-of-tree build, the PROJECT_BINARY_DIR must be
added to the list of include directories. Otherwise, the following build
error occurs:

In file included from
/home/user/iotools/cmos_rw.c:30:
/home/user/iotools/commands.h:23:10: fatal
error: config.h: No such file or directory
   23 | #include "config.h"
      |          ^~~~~~~~~~

Signed-off-by: Yegor Yefremov <yegorslists@googlemail.com>